pyproject.toml
[project]
name = "model-evals-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerebras-cloud-sdk>=1.50.1",
    "duckdb>=1.3.2",
    "fastparquet>=2024.11.0",
    "ipython>=9.5.0",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
